Rename files to follow facebook's conventions
