+++ packages/jest-resolve/src/nodeModulesPaths.js
